% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_db.R
\name{discon_duckdb}
\alias{discon_duckdb}
\title{Ferme le 'pilote' ou 'moteur' duckdb}
\usage{
discon_duckdb(con)
}
\arguments{
\item{con}{Connection}
}
\description{
Ferme une session duckdb en mémoire.
}
\examples{
\dontrun{
con = setup_duckdb() # Exporte con dbdir=':memory:'
discocon(con)
}
}
