% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_hotspots.R
\name{download_ebird_hotspots}
\alias{download_ebird_hotspots}
\title{Télécharge les hotspots eBirds du Québec}
\usage{
download_ebird_hotspots(export_path = ".")
}
\arguments{
\item{export_path}{Chemin d'accès (Défaut : ".", emplacement actuel) pour
déposer les données téléchargées.}
}
\value{
Données téléchargées
}
\description{
Téléchargement de la liste des hotspots eBird avec une clé API de eBird (nécesssite un compte eBird et la clé accessible
au https://ebird.org/data/download). La clé API d'eBird doit être mis dans
le .Renviron avec le nom `EBIRD_API_KEY="<VOTRE_CLE_API>"`.
}
\examples{
\dontrun{
  download_ebird_hotspots()
}

}
