% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_gbif.R
\name{generate_gbif_log}
\alias{generate_gbif_log}
\title{Générer un fichier sommaire de statistiques sur les données GBIF}
\usage{
generate_gbif_log(con = NULL, path_in, path_log = "gb_file_summary.log")
}
\arguments{
\item{con}{Connexion DuckDB (duckdb_connection) valide (`DBIConnection`). Si `NULL`, une
connexion temporaire est initialisée via [setup_duckdb()].}

\item{path_in}{Chemin d'accès (Character) vers le fichier Parquet.}

\item{path_log}{Chaîne de caractères. Chemin d'accès et nom du fichier de sortie (ex: `summary.log`).}
}
\value{
Chaîne de caractères (`character`) contenant le chemin du fichier log, de manière invisible.
}
\description{
Génère un fichier texte (.log) contenant des statistiques descriptives et des
décomptes de lignes basés sur des filtres spécifiques appliqués à un fichier Parquet GBIF.
}
\examples{
\dontrun{
paths <- file.path(
  "data/raw/biodiv/gbif_data/gbif_raw_new.parquet")
# Fichier temporaire
tmp_log_file = tempfile(fileext = ".log")
# Génère le fichier ".log"
generate_gbif_log(path_in = paths, path_log = tmp_log_file)
# Lire le fichier dans la console R
cat(readLines(tmp_log_file), sep = "\n")
}
}
