% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inat_utils.R
\name{get_inat_taxon_id}
\alias{get_inat_taxon_id}
\title{Extraction de l'ID d'un taxon à partir d'un nom scientifique
avec l'API d'iNaturalist}
\usage{
get_inat_taxon_id(scientific_name)
}
\arguments{
\item{scientific_name}{'Character'. Nom scientifique
(genre + épithète spécifique)}
}
\value{
Data.frame avec le résultat de la requête
}
\description{
Requête à l'API \code{(v1)} d'iNaturalist pour obtenir l'ID du taxon et
l'URL.
}
\details{
Les valeurs rapportées proviennent de l'API d'iNaturalist. Si une
correspondance n'est pas exacte, un avertissement est donnée.
}
\examples{
get_inat_taxon_id(scientific_name = "Poecile atricapillus")
}
