% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inat_utils.R
\name{get_taxon_info}
\alias{get_taxon_info}
\title{Obtenir les informations d'un taxon d'iNaturalist}
\usage{
get_taxon_info(nom_latin, attend)
}
\arguments{
\item{nom_latin}{Nom latin d'une espèce}

\item{attend}{Nombre de seconde à attendre. Permet de faire une pause
pour l'API afin de ne pas le surcharger. Ainsi, il est préférable de ne
pas dépasser 60 requêtes par minute et 10 000 requête par jour}
}
\value{
Sortie de la fonction get_inat_taxon_id s'il n'y a pas d'erreur.
}
\description{
Information taxon utilisant \code{get_inat_taxon_id()}
dans un \code{tryCatch}. Permet de faire une pause
pour l'API afin de ne pas le surcharger.
}
\examples{
get_taxon_info(
  nom_latin = "Castor canadensis",
  attend = 1
  )
}
