% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inat_utils.R
\name{get_try_taxon_name}
\alias{get_try_taxon_name}
\title{Test une fonction et attendre}
\usage{
get_try_taxon_name(taxon_id, lang = "french", attend)
}
\arguments{
\item{taxon_id}{ID d'un taxon d'iNaturalist}

\item{lang}{Langue à filtrer dans \code{inat_nom_langue()}
(Default "french")}

\item{attend}{Nombre de seconde à attendre. Permet de faire une pause
pour l'API afin de ne pas le surcharger. Ainsi, il est préférable de ne
pas dépasser 60 requêtes par minute et 10 000 requête par jour.}
}
\value{
Vecteur avec noms français (séparé par ";")
}
\description{
Test une fonction et attendre
}
\details{
Utilise \code{inat_nom_langue()} pour 'tenter' (\code{TryCatch})
lire les données d'iNaturalist.
}
\examples{
# taxon_id = get_taxon_info(
#              taxon_id = "Castor canadensis",
#              attend = 1)$id
fr_nm = get_try_taxon_name(
  taxon_id = 43794,
  attend = .1
)
}
