% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inat_utils.R
\name{inat_nom_langue}
\alias{inat_nom_langue}
\title{Obtenir le nom français d'un taxon depuis l'API d'iNaturalist}
\usage{
inat_nom_langue(taxon_id, collapse_char = ";", lang_filter = "french")
}
\arguments{
\item{taxon_id}{Integer. ID d'un taxon iNaturalist.}

\item{collapse_char}{Caractère de séparation quand plusieurs
noms sont présents dans la sortie (Default: ";")}

\item{lang_filter}{Langue pour filtrer les données (Default: "french")}
}
\value{
Character. Nom français ou NA si absent.
Si multiple noms retournés, les mots sont séparés par des "\code{;}"
}
\description{
Obtenir le nom français d'un taxon depuis l'API d'iNaturalist
}
\examples{
# id_crapaud = get_inat_taxon_id(scientific_name="Anaxyrus americanus")$id
# id_orignal = get_inat_taxon_id(scientific_name="Alces alces")$id
inat_nom_langue(64968) # Crapaud
inat_nom_langue(522193) # Orignal, exemple avec \code{;}
}
