% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inat_utils.R
\name{iNatTry}
\alias{iNatTry}
\title{Tentative de connection à iNaturaliste pour extraire l'information sur
les observation d'espèces}
\usage{
iNatTry(nom_latin, ..., verbose = TRUE)
}
\arguments{
\item{nom_latin}{Nom latin d'une espèce}

\item{...}{Autres arguments de la fonction \code{rinat::get_inat_obs()}}

\item{verbose}{Logique, pour afficher le nom de l'espèce
traitée (Default TRUE)}
}
\value{
Sortie de la fonction \code{rinat::get_inat_obs()} provenant
d'iNaturalist pour obtenir un tableau des observations.
}
\description{
L'utilisation de l'API d'iNaturalist, pour extraire des informations sur
les observations peut parfois causer des erreurs. Dans ces cas, cette
fonction prévient l'arrêt d'exécution en prenant en charge cette erreur.
}
\examples{
\dontrun{
sp.out <- iNatTry(
  nom_latin = "Anaxyrus americanus",
  year = 2025,
  photo_license = "CC0",
  place_id = 6712,  # Canada
  maxresults = 50
)
head(sp.out)
}
}
