% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_db.R
\name{setup_duckdb}
\alias{setup_duckdb}
\title{Initialisation du 'pilote' ou 'moteur' duckdb}
\usage{
setup_duckdb(
  drv = duckdb::duckdb(),
  dbdir = ":memory:",
  extensions = c("spatial"),
  community_extensions = c("h3")
)
}
\arguments{
\item{drv}{'Pilote' pour gestionnaire de BD \code{duckdb::duckdb()}}

\item{dbdir}{Localisation de la BD \code{":memory:"} ou \code{XXX.duckdb}}

\item{extensions}{Vecteur de caractères. Liste des extensions
officielles à installer/charger.}

\item{community_extensions}{Vecteur de caractères. Liste des
extensions de la communauté à installer/charger.}
}
\value{
Connection duckdb
}
\description{
Démarrer une session duckdb en mémoire avec un environnement avec
des extensions spatiales (spatial + h3)
}
\details{
Permet de créer un 'environnement' duckdb prêt à l'analyse avec extensions
utile pour faire des requêtes spatiales
}
\examples{
\dontrun{
require(DBI)
con = setup_duckdb() # Exporte con dbdir=':memory:'

# Exemple de requête dans l'environnement duckdb (et voir les extensions)
DBI::dbGetQuery(
  conn = con,
  statement = "
  SELECT
    extension_name, installed, loaded, description
  FROM
    duckdb_extensions()
  WHERE
    installed = true;" # Filtre ce qui est installé
)
}
}
