% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_gbif_parquet.R
\name{transforme_gbif}
\alias{transforme_gbif}
\title{Transforme données GBIF brutes en fichier spatial PARQUET}
\source{
Voir le ticket GitHub du portail GBIF
\url{https://github.com/gbif/portal-feedback/issues/6570}.
}
\usage{
transforme_gbif(entree, sortie, compression = "zstd", compression_level = 4)
}
\arguments{
\item{entree}{Chemin d'accès (Character) vers le répertoire contenant les
fichiers \code{occurrence.parquet} du GBIF.}

\item{sortie}{Chemin d'accès (Character) du nouveau fichier Parquet
e.g., \code{donnees/gbif_raw_new.parquet}}

\item{compression}{Type de compression pour \verb{COPY() TO (FORMAT PARQUET)}.
(Défaut : "zstd")}

\item{compression_level}{Niveau de compression (Défaut : 4)}
}
\value{
NULL. La fonction exporte directement le fichier
traité sur le disque.
}
\description{
À partir du téléchargement des données de GBIF en au
format \code{SIMPLE_PARQUET} (dossier avec fichiers nommé d'une série de
chiffre comme "003997"),
transforme un répertoire de données brutes GBIF en un unique fichier :
\itemize{
\item Parquet compressé et non partitionné
\item colonne camelCase.
\item colonne géométrique géospatiale (\code{geometry}).
}
}
\details{
L'utilisation de la compression ZSTD (\verb{COMPRESSION 'zstd'}) et
le niveau 4  (\verb{COMPRESSION_LEVEL 4}) permet d'optimiser
l'espace disque tout en conservant la performances de lecture.
}
\note{
Les téléchargements de type \code{SIMPLE_PARQUET} de GBIF génèrent un dossier
contenant des sous-fichiers fragmentés par Apache Spark (e.g., \code{000000}, \code{000001}).
Si l'un de ces fichiers est vide (0 octet, souvent le fichier \code{000000}), DuckDB
échoue lors de la lecture globale. Cette fonction contourne explicitement ce bug
en filtrant les fichiers vides et les fichiers système cachés (commençant par \code{._}).
}
\examples{
\dontrun{
transforme_gbif(
  entree = "occurrence.parquet",
  sortie = "gbif_raw_new.parquet"
)
}
}
