% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat_util.R
\name{xy_convert}
\alias{xy_convert}
\title{Convertir CRS des points (spatiaux) x et y}
\usage{
xy_convert(x, y, crs_to, crs_from = 4326)
}
\arguments{
\item{x}{Numérique. Longitude (ou coordonnée X).}

\item{y}{Numérique. Latitude (ou coordonnée Y).}

\item{crs_to}{CRS final (code EPSG numérique ou objet CRS).}

\item{crs_from}{CRS original (code EPSG numérique ou objet CRS). Par défaut `4326` (WGS84).}
}
\value{
Une matrice (ou vecteur) contenant les coordonnées converties `X` et `Y`.
}
\description{
Transforme les valeurs longitude et latitude
d'un CRS vers un autre CRS.
}
\examples{
xy_convert(
  x = -73.58705319777413,
  y = 45.50351229089908,
  crs_to = 32198,
  crs_from = 4326
)
}
