% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision_algo.R
\name{more_prec}
\alias{more_prec}
\title{Fonction pour augmenter la précision à chaque itération
Utilisation de w (compteur) comme valeur qui va augmenter la précision}
\usage{
more_prec(x)
}
\arguments{
\item{x}{Nombre}
}
\value{
Valeur de sortie selon calcul
}
\description{
Retourne valeur d'une fonction (m1 : 1/x, m2 : exponentielle, m3 :  1/x^2)
}
\examples{
more_prec(1:10)

# Graphique pour montrer la diminution des valeurs de x.
matplot(do.call(cbind, more_prec(1:10)),
  type = 'l',
  xlab = "Indice",
  ylab = "Valeurs", main = "more_prec")
legend("topright",
  legend = c("m1", "m2", "m3"),
  col = 1:3,
  lty = 1:3,
  bty = "o")
}
