% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_grid.R
\name{prep_h3_admin}
\alias{prep_h3_admin}
\title{Ajout des régions admins aux données GBIF}
\usage{
prep_h3_admin(con = NULL, config, res = 10L)
}
\arguments{
\item{con}{Connection à un pilote \code{duckdb()}}

\item{config}{Liste avec les chemins d'accès au minimum
\code{admin_shp} (mis en mémoire avec \code{ST_Read}) et
\code{out_admin_pq} qui exporte un fichier ".parquet"}

\item{res}{"integer": nombre déterminant la résolution H3 e.g., \code{9L}}
}
\value{
Exportation de données \code{out_admin_pq}.
}
\description{
Intersection de polygones administratifs (e.g. municipalités) avec
la grille H3 à fine échelle (e.g., résolution 10L).

L'ajout des H3, aux données administratives, crée une grille standardisée
utile pour en faire référence avec des données GBIF
}
\details{
Connexion à \code{duckdb} et trouver les indices H3 à une résolution \code{res}.
Utiliser \code{transmute} pour garder que les noms '\code{MUS_NM_*}' et
l'indice H3. Copie du fichier selon une préférence
choisie dans le \code{config}.

Pour donner un ordre de grandeur :
Resolution: 4,
"842baa5ffffffff" est un polygone qui
engloble aisément Montréal et Laval (aire ~1930 km2),

Resolution: 7,
"872baa441ffffff" englobe le Mont-Royal (aire ~5.6 km2)

Resolution: 9,
"872baa441ffffff" est plus petit que le Stade-Olympic (aire ~114490 m2)

Resolution: 10, "8a2baa46a50ffff" englobe la salle Wilfrid-Pelletier et
la maison symphonique (aire ~16345 m2)
}
\section{Utilisation potentielle }{

Permet de faire une grille standardiséee (e.g., H3), mais informée par une
variable spatiale (polygones).
Pour faire de la cartographie de points qui
se présenterait en unité spatiale standardisée.
}

\examples{
\dontrun{
# --- Configuration ---
con = setup_duckdb()
folder = "test_path" # Dossier avec les données
paths <- list(
  # Importation
  admin_shp  = file.path(folder, "munic_s.shp"),
  # Exportation
  out_admin_pq  = file.path(folder, "admin_mun.parquet")
)
res = 10L
# --- Calcul---

prep_h3_admin(con, config = paths, res = res)
}
}
